Moved telemetry
